FIX: User ,Aritst, LikeArtist

+++ src/apis/users/users.resolver.ts
@@ -1,5 +1,13 @@
-import { ConflictException, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  NotFoundException,
+  UseGuards,
+} from '@nestjs/common';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { GqlAuthAccessGuard } from 'src/commons/gql-auth.guard';
+import { RoleService } from 'src/commons/role/role.service';
+import { RoleType } from 'src/commons/role/type/role-type';
+import { CurrentUser } from 'src/commons/types/current.type';
 import { CreateUserInput } from './dto/createUserInput';
 import { UpdatePasswordInput } from './dto/updatePasswordInput';
 import { UpdateUserInput } from './dto/updateUserInput';
@@ -8,55 +16,71 @@ import { UsersService } from './users.service';
 
 @Resolver()
 export class UsersResolver {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(
+    private readonly usersService: UsersService,
+    private readonly roleService: RoleService,
+  ) {}
 
   @Mutation(() => User)
   async createUser(@Args('createUserInput') createUserInput: CreateUserInput) {
-    const user = await this.usersService.findOneByEmail({
+    const user_ = await this.usersService.findOneByEmail({
       email: createUserInput.email,
     });
-    if (user) {
+    if (user_) {
       throw new ConflictException('User already exists');
     }
-    return await this.usersService.create({ ...createUserInput });
+    const user = await this.usersService.create({ ...createUserInput });
+    this.roleService.create({
+      userId: user.id,
+      authority: RoleType.USER,
+      artistId: null,
+    });
+    return user;
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Query(() => User)
-  async fetchUser(@Args('userId') userId: string) {
-    const user = await this.usersService.findOne({ userId });
+  async fetchUser(@CurrentUser() currentUser) {
+    const user = await this.usersService.findOne({ userId: currentUser.id });
     if (!user) {
       throw new NotFoundException('User not found');
     }
     return user;
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => User)
   async updateUser(
-    @Args('userId') userId: string,
+    @CurrentUser() currentUser,
     @Args('updateUserInput') updateUserInput: UpdateUserInput,
   ) {
-    const user = await this.usersService.findOne({ userId });
+    const user = await this.usersService.findOne({ userId: currentUser.id });
     if (!user) {
       throw new NotFoundException('User not found');
     }
     return await this.usersService.update({ user, ...updateUserInput });
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
-  async deleteUser(@Args('userId') userId: string) {
-    const user = await this.usersService.findOne({ userId });
+  async deleteUser(@CurrentUser() currentUser) {
+    const user = await this.usersService.findOne({ userId: currentUser.id });
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    return await this.usersService.delete({ userId });
+    return await this.usersService.delete({ userId: currentUser.id });
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Query(() => String)
-  async SendVerificationEmail(@Args('email') email: string) {
+  async SendVerificationEmail(@CurrentUser() currentUser) {
     const authNumber = await this.usersService.createAuthNumber();
-    const template = await this.usersService.getTemplate({ email, authNumber });
+    const template = await this.usersService.getTemplate({
+      email: currentUser.email,
+      authNumber,
+    });
     const result = await this.usersService.sendEmail({
-      email,
+      email: currentUser.email,
       template,
       authNumber,
     });
@@ -64,18 +88,27 @@ export class UsersResolver {
     return result;
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
   async confirmVerificationEmail(
     @Args('authNumber') authNumber: string,
-    @Args('email') email: string,
+    @CurrentUser() currentUser,
   ) {
-    return await this.usersService.verifyAuthNumber({ email, authNumber });
+    return await this.usersService.verifyAuthNumber({
+      email: currentUser.email,
+      authNumber,
+    });
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
   async resetPassword(
+    @CurrentUser() currentUser,
     @Args('updatePasswordInput') updatePasswordInput: UpdatePasswordInput,
   ) {
-    return await this.usersService.updatePassword({ ...updatePasswordInput });
+    return await this.usersService.updatePassword({
+      userId: currentUser.id,
+      ...updatePasswordInput,
+    });
   }
 }
